- used RxDataSource - added dummy data more.

+++ MyTest/MyTest/Domain/Entity/ChatMessage.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct ChatMessage {
-    let messageId: Int64
+    let messageId: Int
     let senderName: String
     let message: String
 }
